+++ sonar-project.properties
@@ -5,5 +5,5 @@ sonar.projectName=mui-rhf-sandbox
 sonar.sources=.
 
 sonar.javascript.lcov.reportPaths=./coverage/lcov.info
-sonar.exclusions=src/tests-related/**,src/**/*.spec.*,src/**/*.test.*,src/pages/**,jest/**,src/api/types/dynamic/**,src/**/index.ts,src/**/*.type.ts
+sonar.exclusions=src/tests-related/**,src/**/*.spec.*,src/**/*.test.*,src/pages/**,jest/**,src/api/dynamic-types/**,src/**/index.ts,src/**/*.type.ts
 sonar.verbose=false
