+++ src/api/dynamic-types/RolesController/getAllRoles.ts
@@ -3,7 +3,7 @@
  * summary: Gets all roles
  */
 
-import { RolesResultDto, ApiResponseDto } from './../api-types';
+import { RolesResultDto, ApiResponseDto } from '../api-types';
 
 export const path = `${process.env.NEXT_PUBLIC_API_URL}/roles`;
