+++ src/api/dynamic-types/SkillsController/getAllSkills.ts
@@ -3,7 +3,7 @@
  * summary: Gets all skills
  */
 
-import { SkillsResultDto, ApiResponseDto } from './../api-types';
+import { SkillsResultDto, ApiResponseDto } from '../api-types';
 
 export const path = `${process.env.NEXT_PUBLIC_API_URL}/skills`;
