+++ src/api/useSkillsQuery.ts
@@ -1,9 +1,9 @@
-import { useAxiosQuery } from './generic/useAxiosQuery';
-import { GetAllRolesError } from './types/dynamic/RolesController/getAllRoles';
+import { GetAllRolesError } from './dynamic-types/RolesController/getAllRoles';
 import {
   GetAllSkillsSuccess,
   path,
-} from './types/dynamic/SkillsController/getAllSkills';
+} from './dynamic-types/SkillsController/getAllSkills';
+import { useAxiosQuery } from './generic/useAxiosQuery';
 
 export const useSkillsQuery = () =>
   useAxiosQuery<GetAllSkillsSuccess, GetAllRolesError>('skills', path, 'GET');
